board: blix-bqx: cfg: optimize DCD command order and values
About 10% of the currently produced boards did not boot
properly on every start. Research and tests showed that
the problem stemmed from a too short time between the
ZQ calibration of the RAM controller, on the one hand,
and calibration of the RAM chips, on the other hand.

While at it optimize some values which were taken from fsl's
sabresd originally and could be finetuned further.

Signed-off-by: Alexander Schartner <aschartner@christ-es.de>
Signed-off-by: Peter Fink <pfink@christ-es.de>